+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 9. Made pronouns be specific to each language, adding more flexibility for translators
 10. Fixed the dynamic favicon when the system theme is set
 11. Made it so that `re: `s in content warnings no longer compound and instead are limited to one automatically placed in
+12. Added a backend setting to set a custom header on responses
 ## v0.13.6
 1. Added one-time invite codes that can be used to manage who can sign up
 2. Made languages with non-ASCII characters load properly on Windows without requiring the unicode characters to be escaped